Add Predicate#intersect, similar to #in, but with ruby's `&` semantics.
This commit adds a predicate that allows checking whether a tuple attribute,
supposed to be collection-valued, intersects some known collection of values.

The predicate is syntactically similar to `#in`, e.g.

    Predicate.intersect(:x, [8, 9])

but has a intersection semantics, as the generated code illustrates:

    !(t[:x] & [8,9]).empty?

The operator only supports being used with an identifier as first argument
and a collection literal as second argument.