Refactor IN expressions to distinguish between literals and opaque.
This commit changes the grammar regarding IN expressions. The right
terms is now a literal or an opaque value, as proper Sexpr nodes.
The `opaque` kind is introduced for Sequel generation with
subqueries (still experimental).

While the AST changes, this is not considered a broken API, since
the factory stays 100% compatible with (non-opaque) right values in
IN expressions.